requirements.txt
aiosqlite==0.19.0
alembic==1.11.1
asyncpg==0.28.0
fastapi[all]==0.99.1
gunicorn==20.1.0
python-decouple==3.6
rich==13.4.2
sqlalchemy-utils==0.41.1
sqlmodel==0.0.8
